Use latest version in auto detection for newer pre releases
